% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate}
\alias{simulate}
\title{Simulate result given data}
\usage{
simulate(input_df)
}
\arguments{
\item{input_df}{data.frame with creative_id and n}
}
\value{
numeric of length 1
}
\description{
Simulate result given data
}
